Require "inherited" keyword for inherited links and props in Schema
EdgeDB Schema now requires explicit "inherited" keyword for links
and properties declared in object types and links which extend
from supertypes with the same link or property.

For example:

    type Parent:
        link foo -> Object

    type Child extending Parent:
        inherited link foo -> Object  # this is OK

    type Child2 extending Parent:
        link foo -> Object  # this is an error now

    type NoParent:
        inherited link foo -> Object  # also an error

The main reason for this change is to make the link inheritance
_explicit_, which benefits readability, but also removes the potential
for a surprise when the link in the parent declaration is removed or
renamed.