Convert to explicitly sized numeric types
We currently have int which is a 64-bit integer, and float, which is a
double precision IEEE754 value.  Postgres supports other numeric
precisions, and it makes sense to expose them.

Naming-wise we follow the modern convention of specifying the bit size
of the type as a suffix, so the new fixed-precision numeric types are:

- `std::int16` -- 16-bit signed integer
- `std::int32` -- 32-bit signed integer
- `std::int64` -- 64-bit signed integer
- `std::float32` -- single presision IEEE754 floating-point type
- `std::float64` -- double presision IEEE754 floating-point type

Additionally `std::decimal` has been renamed to `std::numeric`, and
the following abstract scalar types have been introduced to describe
the numeric type groups:

- `std::anyint` -- any integer type
- `std::anyfloat` -- any floating-point type
- `std::anyreal` -- any non-complex numeric type, includes
  `anyint` and `anyfloat`, as well as `numeric`.

Fixes: #135.