Refactor handling of relation aspects
The 'value' aspect of object relations now represents the identity,
and is equivalent to 'identity' unless the latter is explicitly
overridden.  Previously, the 'value' aspect of object relations was
also used to indicate that the relation can be used to obtain the
values of singular properties.  Now, an explicit 'source' aspect is
used for that purpose.