graphql: Implement reflection of ObjectTypes.
ObjectTypes get reflected as type INTERFACE into GraphQL. Additionally,
non-abstract ObjectTypes get reflected as type OBJECT, but with the
"Type" appended at the end of their name.

This allows the return type of fields/functions to be specified in terms
of INTERFACE types and the actual object __typename to reflect the
OBJECT type of the specific instance.