Stop pulling the "value" aspect unnecessarily
When the compilation does not require the raw value of an expression, it
shouldn't be included in the target list of a corresponding relation.
This is currently an artifact of how the compilation dispatch is
structured, as `dispatch.compile()` is assumed to return a value.  The
solution here is to introduce another dispatch: `dispatch.visit()` which
does not return a value.