Postpone expression cardinality analysis until scope tree is complete
Currently, cardinality inference is happening at the same time as the
expression is compiled.  This is problematic, since there are cases
where the scope tree can be modified by further processing of the
expression, which could affect the inference result.  The most common
such case is expressions in tuples.

To avoid possible inconsistencies, postpone the cardinality inference
until the very end of statement compilation when the scope tree is
complete and final.