Disallow overriding the 'id' property
The value of the `id` property of an object coresponds to object's identity
(and value).  It is both unsafe and unintuitive to allow arbitrary
re-definition of `id`.

```
SELECT Foo ORDER BY Foo;
SELECT Foo ORDER BY Foo.id;
```

The above queries are equivalent by default, but would stop being equivalent
if `id` was re-defined in a view.

Fixes: #214