edgeql: Update the type specification syntax.
Introduce type expressions - syntactical constructs denoting types and
type composition.

Operator `|` produces a union type that is a duck type having the
intersection of all of the links and properties of the component types.
Semantically any object that is an instance of at least one of the
component types is also an instance of the union type.

Operator `&` produces an intersection type that is a duck type having
the union of all of the links and properties of the component types.
Semantically only an object that is an instance of all of the component
types is also an instance of the intersection type.

Syntactically type expressions are general and can be used anywhere
where a type can be used. There may be some exceptions, notably only
base types can be used in `extending` specification.