Fix native serialization of non-singleton links
In native serialization mode non-singleton links are represented as a
record aggregate.  PostgreSQL sometimes "forgets" the structure of an
anonymous tuple type, and so any attempt to access it would fail with
"record type has not been registered".  To combat this, call
BlessTupleDesc() (exposed through the recordext.bless_record() function)
to register the the tuple description in the global cache.