Fix a few cases of missed source path substitution in computables
Current implementation fails to substitute source paths in computable
expressions if the original view shape source is a path, or if a source
path reference is used in a nested view definition, so this:

    SELECT Card {
        owners := (
            SELECT Card.<deck[IS User] {
                friend_count := count(Card.<deck[IS User])
            }
        )
    }

would fail to replace `Card.<deck[IS User]` with `Card.owners` when
evaluating `Card.owners.friend_count`.

Similarly:

    SELECT Card {
        owners := (
            SELECT Foo {
                friend_count := (WITH F := Foo SELECT count(F))
            }
        )
    }

would not replace `Foo` with `Card.owners` in the definition of `F`.

The first case is fixed by recording the original source path id in the
computable source map, and the second case is fixed by correctly stripping
the scope namespaces when looking in the view map.