Untangle handling of scope in views
This round fixes a number of long-standing warts in the handling of
complex scope scenarios in the presence of query iterator
(`FOR` statements), inline expression aliases (`SELECT foo := ...`)
and computables in general.

The main changes here are improving the consistency of scope namespace
handling, as well as explicit disambiguation of computable pointer
paths, so in `SELECT User { name := lower(User.name) }` the shape
element `User.name` and the `User.name` reference in the computable
expression have different path ids.

Inline expression aliases are no longer treated as equivalent to a
`WITH` clause view, because that breaks `FOR` statements
(the iterator alias follows the `WITH` clause logically, hence
the views there cannot refer to the iterator symbol).  The SQL compiler
has also been fixed to handle shaped computables containing a `FOR`
expression.

Finally, view types defined by nested shapes are no longer created
twice, which should improve compiler performance on queries with lots of
nested shapes.