Fix handling of paths that cross a namespace boundary
Schema-level computables are evaluated in a detached scope context.
Practically, this means that all path ids in a computable expression are
given a unique namespace.  However, if a computable expression contains
a `__source__` anchor, the resulting path id has different namespaces
for its head and its tail.  Currently, this information is lost and
path id prefix slices would be invalid in such cases.

The fix is to maintain a reference to an original path whenever a path
extension produces a path with a different namespace.  While at it,
clean up things a bit and replace calls to  `PathId.__getitem__` with
appropriate helper methods.

Fixes: #234.