edgeql: Add JSON funcitonality.
New JSON functions:
- str_to_json
- json_to_str
- json_typeof
- json_array_unpack
- json_object_unpack
- json_get

The `[]` indexing operator is also available for JSON. This operator is
strict and will raise an exception if it is used with a non-existent
index value (this is true for JSON, strings and arrays).

There's a new `array_get` function that doesn't perform the boundary
check on the index value and returns an empty set if there is no such
index.