Switch migration rendering from immediate to buffered approach
Currently, when executed, a migration emits its backend commands
one-by-one, and conditional logic is implemented in the server, which
also has to perform introspection queries to obtain the status of the
backend.

The current approach is obviously inefficient and very tightly coupled
to an active backend connection.  This commit improves the situation by
making the migration execution process to render the necessary backend
commands into a single PL block.  This allows executing the migration
in a single backend roundtrip, as well as decoupling the execution
from script generation.