Add support for casting objects and tuples to std::json
`<json>Object` now always returns a `std::json` value containing
the output representation of the object. That is, the result is
the same as the output of `SELECT Object` in JSON mode, including
the type shape.

Now, casting a non-named tuple to `std::json` returns a JSON array,
and casting a named tuple to `std::json` returns a JSON object.

Casting non-scalar JSON values to scalars is now an error, as well as
casting a non-matching JSON scalar value.

Issue: #251.