Make regular division return floats and add a floor division operator
In another break with tradition, `/` now always performs a true division:
even if the operands are integers, the result is a floating point
approximation of the division result.  The new `//` operator allows
pefforming floor division explicitly.

In this change we largely follow the motivation of Python's PEP 238.
Classic behavior of integer division makes it hard to write polymorphic
functions that work with arbitrary numeric types, especially if one
considers numbers beyond the real plane, as blind casting of one of the
operands may not work as expected.