schema: Add support for spaces + tabs indents.
Indentation must be consistent, so if a mix of tabs and spaces is used,
tabs must come first and shallow indentation must be an exact prefix of
deeper indentation.

It is illegal to use a tab after a space in indentation.

It is illegal to use a line continuation immediately after indentation.
Some other content must appear first on that line.

Whitespace is still ignored in parentheses or in the middle of a line
even if it is extended via line continuation.