Stop interpreting literals into Python values in constants
Currently, the "value" field in various `Constant` AST nodes is a Python
value, which is obtained by interpreting the corresponding EdgeQL
literal.  When rendered back into code, the Python value is then
converted back into an EdgeQL or SQL value.  This approach is inconvenient
and is potentially lossy, so stop doing it and store the actual literal
value as a string, converting to Python only when necessary for
interpretation (such as constant folding).