Implement OPERATOR DDL
The new `{CREATE|UPDATE|DROP} OPERATOR` DDL commands formalize the
process of adding an operator to the schema.  They use the same
parameter and return declaration syntax as functions, although
named-only parameters are obviously not allowed.

The operator DDL is restricted to the `std` module and its only
currently implemented functionality is to alias PostgreSQL
operators to EdgeQL operators.

This approach allows systematic and exhaiustive definition of the
allowed operator space, as well as making it possible to generate proper
polymorphic SQL.