Revert decimal literals back to std::float64
In e793fd37 I changed the interpretation of numerics with a decimal
point to `std::decimal`.  The motivation behind that change was to
adhere to the behavior of PostgreSQL (and the SQL standard
intepretation).  However, it quickly became apparent that this approach,
despite its acceptance in the SQL world, has significant practical
disadvantages.

First and foremost, literals-as-decimals practically requires that
decimals are implicitly castable to floats, otherwise any expression
involving literals and float properties becomes a quagmire of casts.
But casting decimals to floats implicitly loses precision and can be
surprising in other contexts.

Secondly, the vast majority of contemporary programming languages
interpret decimal literals as 64-bit floats.  Many languages have no
built-in concept of arbitrary-precision decimals, which forces a
compromise between an implicit loss of precision or decimal-as-text
representation, both of which are less than ideal.

Thirdly, exact-precision floating point numbers aren't exactly a
necessity in most applications.  Scientific calculations are inherently
performed with errors and tolerances in mind, and majorly prefer the
performance/precision ratio of IEEE floats to that of numerics.
Diligent financial applications tend to avoid fixed- and floating-point
arithmetic as much as possible and compute amounts as integers instead.

For these reasons, numeric literals with a decimal point are now
interpreted as `std::float64`.  Additionally, implicit casts between the
floats and `std::decimal` have been removed, although the assignment cast
is still present.