edgeql: Remove special `INITIAL VALUE` syntax.
Use `SET initial_value` to define the initial value for aggregate
functions. This is just another attribute and should be treated as such.

Regular functions can check if this attribute has been defined and treat
it as an error, but from a syntactical view there is no more special
treatment.

Fixes: #296.