Upgrade flake8 and fix the newly found issues
The new pycodestyle version introduces the new W504 (break after a
binary operator), and E252 (missing whitespace around equal sign in type
annotated function arguments).  Both are reversals of an earlier policy
and, as such, generate an enormous number of errors in the current code,
so I chose to silence instead of fixing them for now.