Remove the need to re-read the schema from backend on DDL
Schema re-introspection on DDL was originally put in place as a
validation mechanism, since at that time the deltas were usually large,
and the overhead of `readschema()` wasn't so bad.  Nowadays we issue
lots of small DDL commands and introspecting the schema after each is
prohibitively expensive, especially now that schema construction takes
more time due to immutability.

Another reason for schema introspection is various schema-to-backend
mappings used to properly resolve the context of errors raised by the
backend etc.  These were historically necessary since we used schema
item names for identification, and names were necessarily mangled in the
backend.  Now, since everything has a stable id, we can just use that
and get rid of lookup mappings, removing the need to introspect the
schema after changes are made.