Rework the way relationship cardinality constraints are expressed
We currently represent relationship cardinality as the `cardinality`
attribute in concrete link and property definitions.  The accepted
values are `'11'` (one-to-one), `'1*'` (one-to-many), `'*1'` (may-to-one),
and `'**'` (many-to-many).  While this is accepted parlour, the
interpretation of the constraints, together with the separate `required`
qualifier produces a fair amount of cognitive load.

In reality, the current monolithic relationship cardinality actually
represents two distinct properties: the cardinality of the target set
when traversing a pointer from a given source (the second '1' or '*'),
and the *exclusivity* constraint on the target (the first '1' or '*').
The first property needs to be specified far more often than the second,
so it makes sense to separate them.

Now, a concrete link or property declaration can include the new target
set cardinality qualifier in the form of the new `single` and `multi`
keywords.  For example:

    type User:
        link friends -> User:
            cardinality: '**'

becomes:

    type User:
        multi link friends -> User

The `single` qualifier is the default, but can still be specified:

    type Issue:
        single link status -> Status

The `required` qualifier can be combined with the `single`/`multi`
qualifier to form a precise target set cardinality requirement:
`single` means zero-or-one, `required single` means `exactly one`,
`multi` means zero-or-more, `required multi` means `one-or-more`.

The second relationship cardinality property is target exclusivity, i.e.
how many sources can relate to a given target, is now represented by the
`exclusive` constraint, which is the current `unique` constraint, but with
a new, more appropriate name.  It is more natural to say that a given
relationship is exclusive, rather than unique. For example, in the
relationship `Person -> owns -> Chair`, "this chair is exclusive to this
person" is better than "the Person/Chair relationship is unique":

    type Person:
        single link owns -> Chair:
	    constraint exclusive

So, to summarize:

Former '11' is now `single` + `constraint exclusive`.
Former '1*' is now `multi` + `constraint exclusive`.
Former '*1' is now just `single` (this is the default).
Former '**' is now just `multi`.

The new qualifier syntax is also significantly more natural to use with
computable links and properties when an explicit cardinality qualifier
needs to be given to override the inferred value:

    type Person:
        multi link items := {'item'}

This now extends to ad-hoc computables in queries as well, so:

    SELECT Person {
        multi items := {'item'}
    }

instead of the awkward and somewhat arcane:

    SELECT Person {
        items := (WITH CARDINALITY '*' SELECT 'item')
    }

The `WITH CARDINALITY '*'` syntax is now entirely unnecessary and has
been removed.

The new syntax also allows expressing the non-zero cardinality
requirement with the `required` qualifier on the computable, but that is
left for a subsequent commit (see #217).

Closes: #303.