edgeql: Add converter functions for multiple data types.
The `to_str` function is meant to be like a cast when called
without the format parameter.

`to_str` usage with `json` scalars is not equivalent to casting, it is
producing a string representation of JSON. The inverse function `to_json`
is therefore parsing a string, not casting either:
- `str_to_json` becomes `to_json`
- `json_to_str` becomes `to_str`

Add various date and time converter functions.