Avoid unnecessary type derivation for the sake of PathId
Current implementation of PathId relies on schema type identity when
comparing paths, which forces us to derive a subtype every time we
generate a PathId for an expression other than a simple path.  This is
inefficient and will cause significant inconvenience with the
schema immutability work.

The solution here is rather simple: use a simple string, representing a
type name, for path comparison.  We already do this for the pointer
portions of the path id.