schema: Reimplement ObjectMapping as ObjectIndex
Currently, `ObjectMapping` stores a key-value dictionary internally.
The key is actually not arbitrary, and is the shortname of the value.
Since names are transient properties, the mapping should not store them
and should refer to the referenced object every time a key is requested.
The new implementation is called an `ObjectIndex`, and there are two
variants: `ObjectIndexByShortname`, which implements the behavior of the
former `ObjectMapping`, and `ObjectIndexByFullname`, which is the same,
but uses the fullname of the referenced object as a key.

This also greatly reduces code duplication between various
`ObjectCollection` subclasses, since the core methods and the underlying
storage are essentially the same.