Stop using derive_copy() when deriving views
Originally, `derive_copy()` was invented as a way to derive an
*independent* variant of a concrete schema item.  The way
`derive_view()` currently uses it is to derive an actual sub-class,
since the original item is written into `bases`.   This causes
unexpected behavior, since the internals of the derived item are
_both_ inherited and copied from parent.

The fix for this is to allow using `.derive()` on concrete schema
items, since there is absolutely no reason to prohibit this anymore.