Stop merging constraints by name
Currently, if several constraints of the same abstract name are declared
on a schema item, these constraints are merged into a single constraint,
where individual constraint expressions are AND-ed.  This approach has
been dicated historically by `ReferencingObject` logic, since
`ObjectMapping` used shortnames as keys.  The downside of this merging is
that it is destructive, i.e. there is no record of the constraint
components that were actually declared on the schema item.  Now that the
`ObjectIndex` has a fullname variant, we can stop merging the
constraints and generate the concrete name from the constraint
expression to disambiguate constraints derived from the same abstract
constraint.