Make schema fully immutable
This is the culmination of the schema immutability effort.  Data for
schema item fields, as well as all schema indexes are now using
`immutable.Map`, and every schema mutation now returns a schema copy.

This change allows us to do things like durable schema snapshots in
transactions, as well as free schema copies.

Co-authored-by: Yury Selivanov (@1st1)