Minimize the number of SQL statements in std bootstrap and cache them
Bundling the entire std bootstrap sequence into a single script block
allows for somewhat faster execution and makes it easier to cache the
script.  Additionally, the std schema pickle is now cached in the data
directory.  I've also reorganized the schema caching to cache into
the 'build/' directory and only in dev mode.