tests: Set default parallelism (-j) to the number of physical cores
It appears that the test suite executes ~10% faster when the number
of test workers is equal to the number of physical cores.

Currently we're using `os.cpu_count()` to get the number of cores.
However, the function returns the number of *logical* cores.  Getting
the number of *physical* cores is not trivial, and it would be easier to
simply add 'psutil' as a dependancy to do that.  However, given that the
number of logical cores is usually 2x of physical ones we simply devide
the result of `os.cpu_count()` by 2.