Restrict usage of abstract scalars and anytype.
Casting into an abstract scalar or anytype is now forbidden.

Technically anytype is a valid type that is the supertype for both
anyscalar and Object. However, it is currently illegal to use it
anywhere outside of a function signature definition.

Anytype is also special in the fact that it doesn't have a corresponding
entry in the schema module (no Type object).