Fix a class of scope tree construction errors
Currently, the following query would return an invalid result due to a
failure of correctly factoring out the `Card` set reference:

    WITH MODULE test
    SELECT <str>count((WITH A := Card SELECT A.owners)) + Card.name

The primary cause is a bug in `ScopeTreeNode.attach_subtree()`, which
currently fails to properly strip nested fence namespaces when looking
for a path node.

The fix for this, however, uncovered a weakness in the computable scope
namespacing system that is used to shield set references in computable
expressions from the computable caller context.  For example, in the
following query:

    WITH MODULE test
    SELECT Card {
        name,
        alice := (SELECT User FILTER User.name = 'Alice')
    } FILTER Card.alice != User;

The `User` reference in the `Card.alice` computable expression must be
treated as separate from the `User` reference in the outer `FILTER`
clause.  Prior to this fix, this worked because the computable
expression had an additional namespace, which shielded the inner `User`
reference.  To keep this working, `computable_ptr_set()` learned to
distinguish references to computable pointers at the site of declaration
(i.e when rendering the view as output), and elsewhere.

This also contains a fix for `DETACHED` expressions forgetting the
computable pointer metadata.