Add test runner markers for more elaborate xfail marking
Currently, we use `@unittest.expectedFailure` regardless of the reason
of the test not passing.  It is useful to a) discriminate between bugs
and not-yet-implemented features; and b) be able to provide a textual
explanation for the test's failure.  This commit adds two decorators:
`edb.tools.test.xfail` and `edb.tools.test.not_implemented` to
implement this.