Merge documentation back to the core edgedb repo.
The initial motivation to keep the documentation in its own repo
was to make it easier to (a) grant commit bits only for docs
modifications; (b) contain the churn of docs updates in a separate
repo; (c) require no copyright assigmnent to make contributions.

However, as we later discovered, having docs separate from the main
code base makes it significantly harder to ensure synchronized updates
to both repos.  Updating documentation properly has been neglected for
the past couple of months.

This commit merges back edgedb-docs repo to the main repo.  The former
repository will be archived.