Subsume ReferencingObject into base Object
Referencing a collection of other objects is actually an
essential trait of any schema object, since we want at least the
ability to associate Attributes with any schema object type.

Hence, it makes little sense to keep the relevant code in
`ReferencingObject`, so move the generic bits into the base `Object`,
and the inheritance-related bits to `InheritingObject`.