Rework schema item attributes
Schema attributes are named values associated with schema items and are
designed to hold arbitrary user-defined schema-level metadata.

There are several shortcomings with the current implementation that
need tweaking.

First issue: attributes can currently be of any primitive type, which
makes it really hard to represent the attribute collection as a
coherent whole without resorting to some string serialization, and that
can go sideways quickly if the attribute type is reasonably complex.
Thus, we stop pretending to support arbitrary types in attributes, and
only support `std::str`.  Complex structures can be stored as JSON and
coerced to a specific EdgeQL type explicitly, so this change has no
practical usability loss.  This also means that abstract attribute
declaration does not require a type specification anymore, although we
may revisit this in the future and allow `std::str` sub-types to be used,
to, for example, allow specifying constraints on attribute values.
Abstract attribute declarations continue to be required for
disambiguation purposes.

Second issue: standard schema item aspects, like link cardinality, are
represented as attributes in the introspection schema. There is no
reason to do it, and this creates confusing duality, since some aspects
are represented as properties and links of the appropriate
`schema::Object` subtype. We stop doing this, and only reserve
`schema::Object.attributes` to things that are true annotations and bear
no effect on the core schema behavior. Currently the only attributes
that satisfy this requirement are `"title"` and `"description"`.

In DDL, the `SET <attribute-name> := <attribute-value>` is changed to
`SET ATTRIBUTE <attribute-name> := <attribute-value>`.

In schema DSL we allow setting attribute values on foreign schema items
using the regular declaration syntax with a qualified name:

    type othermod::Foo:
        attribute attr := 'value'

Issue: #306