Improve polymorphic function and operator resolution
When searching for the appropriate form of a polymorphic callable we
currently look for the form with the minimal total cast distance
between the argument types and the parameter types.  If two or more
forms have the same cast distance, we currently choose whichever was
matched first.  This behavior is obviously incorrect, so this takes some
steps to improve the situation.

In the situation, where multiple form candidates have the same cast
distance, we now attempt to find the form with the minimal total
*type distance* (offset in the ancestor order) between the argument
type and the nearest common ancestor of the argument type and the
candidate callable's parameter type.

For example, consider the following polymorphic function:

    CREATE FUNCTION mean(int64) ...
    CREATE FUNCTION mean(float64) ...

When we call this function with an `int16` argument, both forms would
have the same cast distance of 2 (`int16` -> `int32` -> `int64`, and
`int16` -> `float32` -> `float64`).  However, the nearest common
ancestor of `int16` and `int64`--`anyint`--is closer to `int16` than
`anyreal`, which is the closest common ancestor of `int16` and
`float64`.  Hence, the `int64` form would be chosen in this case.

Finally, if more than one form has the same cast and type distance, and
error is raised, requiring the user to apply explicit casts to
disambiguate the call.