Allow aliasing SQL operators with differing operands
Currently, it's only possible to alias SQL operators with matching
operand types.  Since EdgeQL casting semantics is different from that of
PostgreSQL, we sometimes need to explicitly define operator variants
that are not present in the backend.  The only way to do this currently
is to use the `FROM SQL` declaration.  This makes the operator slower
due to an extra proxy function call.  To remove this price, extend the
`FROM SQL OPERATOR` syntax to allow specifying the operand types of the
backend operator explicitly, like so:

    CREATE OPERATOR ... FROM SQL OPERATOR `>(numeric,numeric)`

Doing so will still create an operator alias with a proxy function, but
in the general case of invocation from a non-polymorphic site, the
backend operator will be used directly with appropriate operand casting.