Allow aliasing operators and functions with different return type
Currently, the return type of an EdgeQL function or operator must match
that of the SQL function or operator it aliases.  It is possible to
workaround by using a `FROM SQL` wrapper, but that is clunky and may
affect performance due to an extra function indirection.  Aggregates do
not have even this workaround.

To fix this, introduce a new std-only `force_return_cast` field.  When
set, it instructs the compiler to automatically insert a cast into the
declared return type.  Convert some std operators and functions which
were relying on wrappers, and fix the return type of certain math
aggregates.