Switch pointer refdict to unqualified name index
We already prohibit referencing links and properties by their
module-qualified name in path expressions.  This takes a step further
and switches the `pointers` refdict to use the unqualified name as the
key, making the unqualified name a unique identifier for the pointer in
the context of its source.

Given this change, the "pointer name" -> "backend column name"
translation is now a nop, which also means that we now restrict the
length of the unqualified pointer name to PostgreSQL's 63 characters.