Remove the dependency of IR on schema
At present, we record schema objects into IR directly, which
means that in order to process the IR, the derived schema must be kept
and passed around.  This also makes any IR serialization impractical, as
the associated schema would need to be serialized as well.

This change moves the necessary information from the schema objects into
the IR directly removing the IR dependency on the schema.