Make everything orderable.
There's a perferctly acceptable natural ordering for `bool` that is also
compatible with the ORDER BY clause.

It makes sense to define ordering for `uuid` and `bytes` as they are not
that different from strings in this perspective.

Objects can be ordered the same way as their id.

We can take the ordering that Postgres seems to define on its jsonb
type as the ordering for our json. This makes it possible to order json
in some stable fashion in the ORDER BY clause.

Add tests for validity of using different scalars in ORDER BY clause.