Add "recursuve" field to operators.
In principle comparison operators for arrays and tuples should have
their behavior derived from the comparison of the elements. We mark
these operators as "recursive" so that the compiler knows how to
validate them.

Only operators with the operands as either all arrays or all tuples can
be recursive. It's illegal to mix recursive and non-recursive versions
of an operator when overloading.