Skip inline target identity optimization for type indirections
When there is a reference to the id property of an object which is linked
to by a link stored inline, we want to route the reference to the inline
attribute.  For example, `Foo.__type__.id` gets resolved to the
`Foo.__type__` column.  However, this optimization must not be applied if
the source is a type indirection, e.g `__type__[IS Array].id`.