Add rudimentary support for the INTROSPECT operator
This adds support for the `INTROSPECT` operator, which is a way to
perform introspection of a given type expression.  Support for
the `TYPEOF` expression is added also, which is a way to perform
static compile-time type inference for an arbitrary expression.

Currently, `INTROSPECT` is limited to the types that are present in the
schema, i.e. introspection of collection types and views is not
supported yet.

I've also converted the uses of `__type__` on scalar expressions to
`INTROSPECT TYPEOF`, as `__type__` on non-object types is slated for
removal.

Issue #219.