Make top-level DML statement return objects instead of count
This reverts DML to its historic behavior, where the result is an
affected set of objects, even if the statement is top-level.  The prior
change that made this return a count was a dubious attempt of optimization,
so that the output of DML can be discarded.  However, this case should
be solved by a protocol context rather than this hardcoded behavior.