Add coverage support to the `edb test` command
It's now possible to measure test coverage with the `edb test` command.

Pass `--cov` (allowed to be specified multiple times) to measure
coverage of a particular edb package, e.g.:

* edb test --cov=edb
* edb test --cov=edb.edgeql.compiler
* edb test --cov=edb.edgeql.compiler --cov=edb.pgsql.compiler

The results will be printed to the stdout when tests are completed.
They will also be saved to the '.coverage' file, so that the 'coverage'
command line tooling can be used for further inspection.