Implement casting of arbitrary arrays to JSON, fix assorted issues
Currently, when casting to std::json, only an explicit array literal
is supported.  Expand the casting support to allow for any array value
by applying the cast to the array element-wise.

Along the way, several defects have surfaced in JSON output
serialization, and the `array_unpack' and `array_enumerate` functions,
specifically when unpacking tuples.  Fix these by a more thorough
application of the `RangeFunction` SQL syntax and careful management of
coldeflist.  A side-effect of this fix is that the order of elements in
the `array_enumerate` result tuple is now reversed, with the index now
the first element, so the resemblance to similar constructs in other
languages is closer.