Remove support for __type__ pseudo-link on primitive types
Historically, scalar types and collections supported `.__type__` as a
way to determine the type of an expression.  Now that the `TYPEOF` and
`INTROSPECT` operators have been implemented, this hack is no longer
necessary.

This commit also fixes the `IS` operator on operands of primitive types.

`IS` on collections containing object types remains unsupported, as
implementing that requires a non-trivial transform and the use cases do
not seem clear enough.

See also: #219.
Closes: #161.