Make assert_query_result() test both JSON & Binary outputs
The function now first runs query in a transaction with the
'fetch_json()' method.  If the returned result matches the
provided pattern, the query is run for the second time via
the `fetch()` method.  The output is then converted to a
JSON-like structure to test it against the provided pattern.