Make sure object value ref and explicit id prop ref are same (try 2)
In 2a0e0941f, I made an attempt to ensure that subsequent refs to either
the object value or to object.id property value always result in the
same ref.  There seems to be more modes when this breaks, so make a more
aggressive effort at synchronizing these.

Fixes test_edgeql_select_setops_13 in binary mode.