Fix computable cardinality inference in the presence of LIMIT
The LIMIT and OFFSET clauses put additional scope fencing around, which
is causing issues with correct computable cardinality inference,
including that of the implicit __tid__.

One culprit is an old LIMIT-specific scope hack, which is now
unnecessary, and, evidently, harmful.  Another is the fact that the
scope tree path namespace is sometimes set up incorrectly with duplicate
namespaces in descendant nodes, so prevent that from happening.