Generate different IDs for shapes with and without implicit fields
We always inject object id to the output data, whether it was explicitly
requested or not.  This means that the following two queries have
identical data output:

* `SELECT Object {name};`
* `SELECT Object {id, name};`

The only difference is that the second query specifies that the 'id'
field is explicit in its type description metadata.  We now mix this
information in the query output type UUID, so that clients can correctly
cache object codecs (and see the difference between the two queries
above.)