Update on target delete triggers.
Rename the `set empty` option to `allow`. This is intended as a
complementary to the default `restrict`. Plus the naming is less
misleading for multilinks.

Drop the `set default` option as it's not currently implemented

Document all the different behavior options for when a link target is
deleted without explicitly updating the link.