Fix an issue with repl mishandling `;` or lack of it.
The `;` is supposed to be a separator, so multiple `;` in a row should
have no effect. REPL should not be attempting to execute a statement
that's just a `;`.

Also, the `;` is not required for a single statement, so REPL should be
capable of handling that case too (terminating input with alt+enter).