Fix <array<tuple<...>>> casts and case of set of tuples returning func
Array casts should fall back to the generic cast if no direct element
cast was found, since tuples have no explicit casts in the schema.

Another issue fixed here is a case when a set returning function returns
a tuple of one element, which is currently erroneously treated as a
non-tuple value.