Make type operation use consistent.
The way link targets were specified in eschema was different from DDL
for multiple type links.

DDL and EdgeQL operator `IS` make use of type union operator `|` instead
of a `,` to specify multiple types. This is now the case for eschema,
too.

Union types can now be specified as link targets in eschema and in DDL.
They can also be used as the RHS of the `IS` operator.

Union types do not yet implement the "common" property/link logic.

Make the `__type__` link point to `ObjectType` instead of `Type`.

Documentation for types, links, and type operators and introspection
has been updated.