Streamline handling of type indirections in paths
The current situation with type indirections in paths (`[IS Type]`) is
fairly messy, mostly for historical reasons.  Originally, the `[IS ..]`
syntax (and its predecessors) were used to specify a concrete target
type for a link in a path, e.g. `Issue.references[IS File]`.
Subsequently, this was extended to arbitrary expressions, not just link
paths, so `Named[IS Text].body` became possible.  Finally, shapes also
have a dedicated syntax to specify either the source type indirection
(`{..., [IS Type].ptr}`), or the target type indirection
(`{..., link: Type }`).  All of these mechanisms are currently handled
slightly differently, and are specified in two AST places: `Ptr.target`
for shapes and link paths, and `TypeFilter` for all other expressions.
The `Ptr.target` mechanism also relies on a piece of complicated logic
in `Source.resolve_pointer()` to derive ad-hoc link instances that
satisfy a given path expression.

None of the existing complexity is really necessary, as the type
indirection is already treated as just another form of path indirection
(other being object pointer and link property references).  Given that
fact, all above variants can be represented with an indirection step, so

    Expr[IS Type] --> Path(steps = [Expr, TypeIndirection(Type)])

    Obj.link[IS Type] -> Path(steps = [Obj, Ptr(link),
                                       TypeIndirection(Type)]

    Obj { [IS Type].foo } is sugar for Obj { foo := Obj[IS Type].foo }

    Obj { foo: Type } is sugar for Obj { foo := Obj.foo[IS Type] }

This cleanup also fixes bugs inherent to the old construction,
so expressions like `Obj.link[IS Foo][IS Bar]` and
'Obj { [IS Type].foo: Target }` are now handled correctly.